Add build script for OpenMPI base container
According to Apptainer's Hybrid/Bind MPI models, we need to build an
OpenMPI that's compatible with the outside world inside our
container. While we're here, we also install the rest of the build
dependencies, including MKL, because this base container should be
relatively static compared to the Firedrake build container, which
could be rebuilt more often.